requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.2
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
boto3==1.40.55
botocore==1.40.55
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.13
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.2
Faker==37.1.1
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
msgpack==1.1.2
mysqlclient==2.2.7
oauthlib==3.3.1
pillow==11.3.0
platformdirs==4.3.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==22.3
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
starlette==0.46.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.30.0
zope.interface==8.0.1
